Force the ECS service to depend on the load balancer
This is because the ECS service tries to immediately attach to its target
group, but the target group is not attached yet to the load balancer
(because the load balancer takes longer to create than the ECS stuff). By
making the ECS service depend on the listener, we force the ECS service to be
created after the load balancer, the listener and the target group are all
created and ready to run.