Fix Docker images
This is to ensure the customer won't shoot themselves in the foot. When
updating a stack from the CloudFormation console, previous parameter
values are used instead of using default values specified in the
CloudFormation template. So if we specify the Docker images as
parameters, the customer will need to update every single one of them
with the right tag, which is unlikely to end up well.